correct sync of composer files for patches

See merge request jtlshop/shop4!2538

+++ build/scripts/build.sh
@@ -257,20 +257,20 @@ build_add_files_to_patch_dir()
     rsync -R admin/includes/shopmd5files/${APPLICATION_VERSION_STR}.csv ${PATCH_DIR};
     rsync -R includes/defines_inc.php ${PATCH_DIR};
 
-    if [[ -f "${PATCH_DIR}/includes/composer.json" ]]; then
+    if [[ -f "${PATCH_DIR}/includes/composer.lock" ]]; then
         mkdir /tmp/composer_${PATCH_VERSION};
         mkdir /tmp/composer_${PATCH_VERSION}/includes;
-        touch /tmp/composer_${PATCH_VERSION}/includes/composer.json;
         git show ${PATCH_VERSION}:includes/composer.json > /tmp/composer_${PATCH_VERSION}/includes/composer.json;
+        git show ${PATCH_VERSION}:includes/composer.lock > /tmp/composer_${PATCH_VERSION}/includes/composer.lock;
         composer install --no-dev -q -d /tmp/composer_${PATCH_VERSION}/includes;
 
         while read -r line;
         do
             path=$(echo "${line}" | grep "^Files.*differ$" | sed 's/^Files .* and \(.*\) differ$/\1/');
             if [[ -z "${path}" ]]; then
-                filename=$(echo "${line}" | grep "^Only in includes\/vendor: .*$" | sed 's/^Only in includes\/vendor: \(.*\)$/\1/');
-                if [[ ! -z "${filename}" ]]; then
-                    path="includes/vendor/${filename}";
+                filename=$(echo "${line}" | grep "^Only in includes\/vendor.*: .*$" | sed 's/^Only in \(includes\/vendor[\/]*.*\): \(.*\)$/\1\/\2/');
+                if [[ ! -z "${filename}" ]] && [[ -f ${filename} ]]; then
+                    path="${filename}";
                     rsync -Ra -f"+ *" ${path} ${PATCH_DIR};
                 fi
             else
